+++ sereto/report.py
@@ -7,12 +7,12 @@ from sereto.models.version import ProjectVersion
 
 
 @validate_call
-def render_report_to_tex(project_path: DirectoryPath, config: Config, version: ProjectVersion) -> str:
+def render_report_to_tex(project_path: DirectoryPath, template: str, config: Config, version: ProjectVersion) -> str:
     """Render the report (top-level document) to TeX format."""
     # Construct path to report template
-    template = project_path / "layouts/report.tex.j2"
-    if not template.is_file():
-        raise SeretoPathError(f"template not found: '{template}'")
+    template_path = project_path / f"layouts/{template}.tex.j2"
+    if not template_path.is_file():
+        raise SeretoPathError(f"template not found: '{template_path}'")
 
     # Render Jinja2 template
     return render_jinja2(
@@ -22,7 +22,7 @@ def render_report_to_tex(project_path: DirectoryPath, config: Config, version: P
             project_path / "includes",
             project_path,
         ],
-        file=template,
+        file=template_path,
         vars={
             "c": config.at_version(version=version),
             "config": config,